# Conditionally generate fp64 viewport uniforms