--- a/test/src/core/shaderlib/project/viewport-uniforms.spec.js
+++ b/test/src/core/shaderlib/project/viewport-uniforms.spec.js
@@ -19,7 +19,6 @@
 // THE SOFTWARE.
 
 import test from 'tape-catch';
-import {Matrix4} from 'luma.gl';
 
 import {COORDINATE_SYSTEM, Viewport, WebMercatorViewport} from 'deck.gl';
 import {getUniformsFromViewport} from 'deck.gl/core/shaderlib/project/viewport-uniforms';
@@ -54,12 +53,13 @@ const UNIFORMS = {
   project_uModelMatrix: Array,
   project_uViewProjectionMatrix: Array,
 
-  // 64 bit support
-  project_uViewProjectionMatrixFP64: Array,
-
   // This is for lighting calculations
-  project_uCameraPosition: Array,
+  project_uCameraPosition: Array
+};
 
+// 64 bit support
+const UNIFORMS_64 = {
+  project_uViewProjectionMatrixFP64: Array,
   project64_uViewProjectionMatrix: Array,
   project64_uScale: Number
 };
@@ -77,8 +77,10 @@ const DEPRECATED_UNIFORMS = {
   projectionScale: Number, // This is the mercator scale (2 ** zoom)
   viewportSize: Array,
   devicePixelRatio: Number,
-  cameraPos: Array,
+  cameraPos: Array
+};
 
+const DEPRECATED_UNIFORMS_64 = {
   projectionFP64: Array,
   projectionScaleFP64: Array
 };
@@ -97,6 +99,14 @@ test('project#getUniformsFromViewport#shader module style uniforms', t => {
 
   for (const uniform in UNIFORMS) {
     t.ok(uniforms[uniform] !== undefined, `Returned ${uniform}`);
+  }
+  for (const uniform in UNIFORMS_64) {
+    t.ok(uniforms[uniform] === undefined, `Should not return ${uniform}`);
+  }
+
+  uniforms = getUniformsFromViewport({viewport, fp64: true});
+  for (const uniform in UNIFORMS_64) {
+    t.ok(uniforms[uniform] !== undefined, `Return ${uniform}`);
   }
 
   uniforms = getUniformsFromViewport({
@@ -117,10 +127,17 @@ test('preoject#getUniformsFromViewport#deprecated uniforms', t => {
   for (const uniform in DEPRECATED_UNIFORMS) {
     t.ok(uniforms[uniform] !== undefined, `Returned deprecated ${uniform}`);
   }
+  for (const uniform in DEPRECATED_UNIFORMS_64) {
+    t.ok(uniforms[uniform] === undefined, `Should not return deprecated ${uniform}`);
+  }
+
+  uniforms = getUniformsFromViewport({viewport, fp64: true});
+  for (const uniform in DEPRECATED_UNIFORMS_64) {
+    t.ok(uniforms[uniform] !== undefined, `Return deprecated ${uniform}`);
+  }
 
   t.ok(uniforms.devicePixelRatio > 0, 'Returned devicePixelRatio');
-  t.ok((uniforms.projectionMatrix instanceof Float32Array) ||
-    (uniforms.projectionMatrix instanceof Matrix4), 'Returned projectionMatrix');
+  t.is(uniforms.projectionMatrix.length, 16, 'Returned projectionMatrix');
 
   uniforms = getUniformsFromViewport({
     viewport,