During profiling of an internal app, `setUniforms` and `fp64ifyMatrix4` stand out among the top most expensive operations. In reality, we do not have to generate the fp64 uniforms when using 32bit projection. 

Changes in this diff:
- Only generate fp64 viewport uniforms if `fp64` flag is on
- Remove duplicate call to `fp64ifyMatrix4(viewProjectionMatrix)`
- Remove unnecessary creation of new typed arrays
- Reduce small arrays generated during `fp64ify`

In terms of implementation, I do not think checking `fp64` is the right approach, as layers could use the `project64` module without the `fp64` flag. I suggest that in luma.gl, when a module specifies `dependencies`, its `getUniforms` callback should be able to access the uniforms generated by its dependencies. `project64` will then be able to leverage the `viewProjectionMatrix` generated by the `project` module.

Tested: test-node, test-browser, layer-browser