@Pessimistress 2017-12-07
**Benchmark in browser (Chrome 62)**

Before:

| Test | Result |
| ---- | ------ |
| getUniformsFromViewport#LNGLAT | 30.5k iterations/s (0.33s) |
| getUniformsFromViewport#METER_OFFSETS | 22.7k iterations/s (0.44s) |
| getUniformsFromViewport#LNGLAT - FP64 | 35.5k iterations/s (0.28s) |
| getUniformsFromViewport#METER_OFFSETS - FP64 | 22.7k iterations/s (0.44s) |

After:

| Test | Result |
| ---- | ------ |
| getUniformsFromViewport#LNGLAT | 1.47M iterations/s (0.68s) |
| getUniformsFromViewport#METER_OFFSETS | 282k iterations/s (0.35s) |
| getUniformsFromViewport#LNGLAT - FP64 | 129k iterations/s (0.78s) |
| getUniformsFromViewport#METER_OFFSETS - FP64 | 96.7k iterations/s (1.03s) |